docs: write for a reviewer who has never seen this branch
A polish pass over the crate ahead of review. No behaviour changes: 139
tests before and after, and the only non-comment edits are the module
split and its import wiring.

Comments now explain the code rather than its edit history. Twenty-seven
of them narrated what a previous revision did — "this used to be", "the
old body was", internal task ids like A20 / A30 / R2 / R5a / Iter 7, a
link to an issue in a fork nobody upstream can read. That is context a
reader here does not have and cannot get, and it crowds out the part that
does help: what the TypeScript runtime does and why this has to match. Each
one is rewritten to state the rule instead of the diff.

Three doc links pointed at files that do not exist on `main`: the Rust
codegen user guide, `compiler/rust-passes-types.ss`, and a fork-local
research note. All three describe the compiler backend, which is a later
contribution — clicking them from this PR gives a 404. Replaced with prose
that stands on its own. A garbled sentence in the README (two edits
collided inside it) is fixed too.

`adts.rs` was 1088 lines doing three unrelated jobs. Split along the seams
that were already there:

  adts.rs               1088  ->    63   Counter, serialize_contract_state
  decode.rs                        560   width-typed ledger decoders
  decode_field_repr.rs             540   the alignment walk for composites

Every file in the crate is now under 370 lines of non-test code, which
means no single review sitting has to hold more than one idea at a time.
The alignment walk in particular is the hardest thing here to review, and
it now has its own file and a header explaining why an atom is not one
field element.

The README gains a section comparing this crate with the TypeScript
runtime, which is the question a reviewer will actually have. It is built
from reading `runtime/src/`, and the numbers in it are checked: 60 `throw`
sites, 12 destructive `shift`/`splice` calls in `fromValue`, and four
distinct Compact types — `Field`, `Uint<0..n>`, `Secp256k1Base`,
`Secp256k1Scalar` — all declared `CompactType<bigint>`, so the host type
cannot tell them apart.

The sharpest example is witnesses, because both backends emit code for it.
`print-typescript.ss` generates `typeof(witnesses.foo) !== 'function'`
guards, one per witness plus one for the bag: a contract with n witnesses
ships n+1 checks whose whole job is to verify at run time what a type
system verifies at compile time. The Rust backend emits a `Witnesses<PS>`
trait and no checks — and the trait is strictly stronger, since
`typeof x === 'function'` says nothing about arity or return type.

Also documents where types do *not* help, because that is where the real
bugs were: encoding rules. `hashToCurve` hashes the field-aligned value
rather than the Rust `FieldRepr`; a Compact type's domain is not its
storage width; `JubjubPoint` is a coordinate pair rather than a validated
group element. Each has a test.

Gates: cargo fmt, clippy -D warnings, 139 tests, and
RUSTDOCFLAGS=-Dwarnings cargo doc all pass. Root workspace and compiler
untouched; runtime-rs/Cargo.lock unchanged.
