+++ runtime-rs/src/op_builder.rs
@@ -239,7 +239,7 @@ impl<D: DB> OpProgramGather<D> {
     /// `size` — replace the top-of-stack container (Map / Set / List /
     /// Array) with a `Cell(u64)` holding its element count. Emitted by the
     /// read-no-arg ADT ops `Set.size`, `Set.isEmpty`, `Map.size`,
-    /// `Map.isEmpty` (A20 — read-no-arg adt-op vm-code lowering).
+    /// `Map.isEmpty`.
     pub fn size(mut self) -> Self {
         self.ops.push(Op::Size);
         self
@@ -248,7 +248,7 @@ impl<D: DB> OpProgramGather<D> {
     /// `type` — replace the top-of-stack `StateValue` with a `Cell` whose
     /// 1-byte payload tags the value's shape (Null=0, Cell=1, Map=2, …).
     /// Emitted by `List.isEmpty` to detect the `Null` sentinel at the head
-    /// of an empty cons list (A20 — read-no-arg adt-op vm-code lowering).
+    /// of an empty cons list.
     pub fn type_(mut self) -> Self {
         self.ops.push(Op::Type);
         self
